HRMW-1059 - Refactor ProjectEmployee service and controller for pagination and flexibility
The ProjectEmployeeService has been refactored to simplify the retrieval process of employees in a project and support pagination. Similarly, the ProjectEmployeeController code has been refactored to provide flexibility for finding and adding employees to a project, providing different behavior based on user specifications for adding either active employees or specific employees to a project.